Added check for an empty file in SignFile. Removed problematic and unnecessary reinterpret cast. Randomised the sie of string that gets signed in test for Sign. Changed the return of Size in safe queue to a size_t.

+++ include/maidsafe/common/safe_queue.h
@@ -57,7 +57,7 @@ class SafeQueue {
     return queue_.empty();
   }
 
-  unsigned Size() const {
+  size_t Size() const {
     std::lock_guard<std::mutex> lock(mutex_);
     return queue_.size();
   }
